requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cssselect==1.2.0
cycler==0.11.0
docutils==0.19
Flask==2.2.3
flask-paginate==2022.1.8
fonttools==4.39.2
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kaleido==0.2.1
kiwisolver==1.4.4
lockfile==0.12.2
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parsel==1.7.0
Pillow==9.4.0
plotly==5.13.1
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-daemon==3.0.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rsa==4.9
schedule==1.1.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selectorlib==0.16.0
six==1.16.0
tenacity==8.2.2
threadpoolctl==3.1.0
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.15
w3lib==2.1.1
Werkzeug==2.2.3
wordcloud==1.8.2.2
